Allow user-customisable backing storage.
Previously the backing storage was fixed to u64, which probably isn't optimal
for non-64 bit machines. The default is now to use usize storage as the backing,
which is probably the best option in nearly all cases: but, in case someone
knows something more (e.g. a machine where usize is 64-bit, but 128-bit integer
operations are as fast as 64-bit operations), we allow them to override this. In
general, however, this has no effect on the API.